userspace: fix up play_sound syscall

+++ userspace/vlibc/include/vmwos.h
@@ -20,5 +20,5 @@ int vmwos_random(uint32_t *buffer);
 int vmwos_framebuffer_load(int x, int y, int depth, char *fb);
 void *vmwos_malloc(uint32_t size);
 int vmwos_core_poke(uint32_t which);
-int vmwos_play_sound(uint8_t *buffer, uint32_t length, uint32_t repeat);
+int vmwos_play_sound(uint32_t *buffer, uint32_t length, uint32_t repeat);
